refactor: :recycle: シーンの制御をFSDのものに置き換えました
