+++ packages/app/src/features/scene/scene.store.ts
@@ -1,0 +1,4 @@
+import { atom } from "jotai";
+import type { Scene } from "./scene.model";
+
+export const sceneAtom = atom<Scene>("mode-select");
